+++ src/lib.rs
@@ -1,7 +1,7 @@
 pub mod common;
 pub mod compute;
-pub mod frost;
 pub mod schnorr;
 pub mod traits;
 pub mod util;
+pub mod v1;
 pub mod vss;
